requirements.txt
ufal.udpipe==1.2.0.3
allennlp==2.10.0
allennlp-models==2.10.0
logzero
